introduce time control

+++ abstractions/Models/NewGameModel.cs
@@ -5,8 +5,8 @@ public class NewGameModel
     public string GameId { get;set; }
     public string Engine { get; set; }
     public int PowerCap { get; set; } // POWER CAP IS PERCENTAGE VALUE POWER = (MAX WATTS - MIN WATTS) * (POWER CAP / 100)) + MIN WATTS
-    public int TimeControl { get; set; }
-    public int TimeIncrement { get; set; }
+    public long TimeControl { get; set; } // MS
+    public long TimeIncrement { get; set; } // MS
     public string InitialPosition { get; set; }
     public string PowerCapColor { get; set; }
 }
